+++ CHANGELOG.rst
@@ -1,5 +1,12 @@
 Changes
 =======
+Version 0.10.7
+-------------
+**2024-4-2**
+
+* Added ``correct_countdown_lists`` method, which applies to corrections to
+  ``list`` field of countdown events for sessions containing unityEPL-FR bug.
+
 Version 0.10.6
 -------------
 **2024-3-8**